+++ src/Validation/dtos/index.ts
@@ -17,3 +17,5 @@ export { IValidationIsstringDto } from "./validation.isstring.dto";
 export { IValidationTrimDto } from "./validation.trim.dto";
 export { IValidationReplaceDto } from "./validation.replace.dto";
 export { IValidationIsjwtDto } from "./validation.isjwt.dto";
+export { IValidationIsStrongPasswordDto } from "./validation.isstrongpassword.dto"
+export { IValidationIsEmailDto } from "./validation.isemail.dto"